Add initial code for server
This is one potential way of doing the core server code, where it only
uses a single UDP socket for all sending and receiving. The server spins
up a thread that will listen forever on the socket and then run
registered handlers in their own threads based on a "type" field in the
JSON-formatted messages that it receives (so this system assumes all
messages are valid JSON).

Ideally, this means that all of the database logic and response message
sending code can just be put into the handler functions.